Address feedback on flows models refactor
