--- src/flows/models.py
@@ -1,550 +1,0 @@
-import functools
-import json
-import operator
-from types import SimpleNamespace
-from typing import List
-
-from django.db import models
-from django.utils.functional import cached_property
-
-from flows.consts import OPERATOR_MAP, FlowActionChoices
-from flows.flow_event import FlowEvent
-
-# A constant encapsulating all conditional actions.
-CONDITIONAL_ACTIONS = {
-    FlowActionChoices.EVALUATE_EQUALS,
-    FlowActionChoices.EVALUATE_NOT_EQUALS,
-    FlowActionChoices.EVALUATE_LESS_THAN,
-    FlowActionChoices.EVALUATE_GREATER_THAN,
-    FlowActionChoices.EVALUATE_LESS_THAN_OR_EQUALS,
-    FlowActionChoices.EVALUATE_GREATER_THAN_OR_EQUALS,
-}
-
-
-class Event(models.Model):
-    """
-    Represents an event type that triggers can listen for or emit.
-    """
-
-    key = models.CharField(
-        max_length=50, primary_key=True, help_text="Unique identifier for the event."
-    )
-    label = models.CharField(
-        max_length=255, help_text="Human-readable label for the event."
-    )
-
-    def __str__(self):
-        return self.label
-
-
-class FlowDefinition(models.Model):
-    """
-    Represents a reusable definition for a flow, consisting of multiple steps.
-    """
-
-    name = models.CharField(max_length=255, unique=True)
-    description = models.TextField(blank=True, null=True)
-
-    def __str__(self):
-        return self.name
-
-    @staticmethod
-    def emit_event_definition(event_name: str) -> "FlowDefinition":
-        """Create an unsaved FlowDefinition that emits ``event_name``."""
-        flow_def = FlowDefinition(name=f"_emit_{event_name}")
-        step = FlowStepDefinition(
-            flow=flow_def,
-            action=FlowActionChoices.EMIT_FLOW_EVENT,
-            variable_name="emit_event",
-            parameters={"event_type": event_name},
-        )
-        flow_def._unsaved_steps = [step]
-        flow_def.steps = SimpleNamespace(all=lambda: flow_def._unsaved_steps)
-        return flow_def
-
-
-class FlowStepDefinition(models.Model):
-    """
-    Represents a single step in a flow definition.
-
-    The 'variable_name' field serves as a generic reference. Its meaning depends on the
-     action:
-      - For conditional steps (evaluate_equals, etc.), it names the flow variable to
-      test.
-      - For SET_CONTEXT_VALUE, it indicates which context variable to set.
-      - For CALL_SERVICE_FUNCTION, it names the service function (or could be used to
-      look it up).
-
-    The 'parameters' JSONField stores extra parameters for the step. For a condition,
-    it might be:
-      { "value": "10" }
-    For a service function, it might be a mapping of keyword arguments, e.g.,
-      { "modifier": "5", "result_variable": "attack_result" }
-    """
-
-    flow = models.ForeignKey(
-        FlowDefinition,
-        related_name="steps",
-        on_delete=models.CASCADE,
-        help_text="The flow this step belongs to.",
-    )
-    parent = models.ForeignKey(
-        "self",
-        null=True,
-        blank=True,
-        related_name="children",
-        on_delete=models.CASCADE,
-        help_text="The parent step of this step.",
-    )
-    action = models.CharField(
-        max_length=50,
-        choices=FlowActionChoices.choices,
-        help_text="The action this step performs.",
-    )
-    variable_name = models.CharField(
-        max_length=255,
-        blank=True,
-        help_text="The key whose meaning depends on the action: for condition steps, "
-        "the flow variable to evaluate; for service functions, the name of "
-        "the service function or target context variable.",
-    )
-    parameters = models.JSONField(
-        blank=True,
-        help_text="A JSON object with parameters. For conditions, should include a key "
-        "'value' (the value to compare against). For service functions, it’s "
-        "a mapping of kwargs (and optionally 'result_variable' to store the "
-        "output).",
-    )
-
-    def execute(self, flow_execution):
-        """
-        Executes this step within the given FlowExecution context.
-
-        - For conditional steps, retrieves the flow variable specified by 'variable_name',
-          casts the 'value' in parameters to the type of that variable, and compares using
-          the operator defined by the action. If the condition passes, returns the next child;
-          otherwise, returns the next sibling.
-        - For SET_CONTEXT_VALUE/MODIFY_CONTEXT_VALUE, updates context data.
-        - For CALL_SERVICE_FUNCTION, calls the service function and stores any result.
-        - For EMIT_FLOW_EVENT, creates and stores a FlowEvent in the context.
-
-        Returns the next FlowStepDefinition to execute, or None if the flow is complete.
-        """
-        if self.action in CONDITIONAL_ACTIONS:
-            condition_passed = self._execute_conditional(flow_execution)
-            if condition_passed:
-                return flow_execution.get_next_child(self)
-            else:
-                return flow_execution.get_next_sibling(self)
-        elif self.action == FlowActionChoices.SET_CONTEXT_VALUE:
-            return self._execute_set_context_value(flow_execution)
-        elif self.action == FlowActionChoices.MODIFY_CONTEXT_VALUE:
-            return self._execute_modify_context_value(flow_execution)
-        elif self.action == FlowActionChoices.CALL_SERVICE_FUNCTION:
-            return self._execute_call_service_function(flow_execution)
-        elif self.action == FlowActionChoices.EMIT_FLOW_EVENT:
-            return self._execute_emit_flow_event(flow_execution)
-        elif self.action == FlowActionChoices.EMIT_FLOW_EVENT_FOR_EACH:
-            return self._execute_emit_flow_event_for_each(flow_execution)
-        else:
-            return flow_execution.get_next_child(self)
-
-    def _execute_conditional(self, flow_execution):
-        """
-        Executes a conditional step by comparing a flow variable with a provided value.
-        Returns True if the condition passes, False otherwise.
-        """
-        left_value = flow_execution.get_variable(self.variable_name)
-        op_func = OPERATOR_MAP[self.action]
-        comp_raw = self.parameters.get("value")
-        right_value = type(left_value)(comp_raw)
-        return op_func(left_value, right_value)
-
-    def _execute_set_context_value(self, flow_execution):
-        """
-        variable_name - the *flow variable* holding an object pk
-        parameters    - {
-                            "attribute": "<field in ObjectState>",
-                            "value":     <any JSON-serialisable literal>
-                        }
-        """
-        object_pk = flow_execution.get_variable(self.variable_name)
-        if object_pk is None:
-            raise RuntimeError(
-                f"Flow variable '{self.variable_name}' is undefined – "
-                "cannot set context value."
-            )
-
-        attribute_name = self.parameters["attribute"]
-        literal_value = self.parameters["value"]
-
-        flow_execution.context.set_context_value(
-            key=object_pk,
-            attribute=attribute_name,
-            value=literal_value,
-        )
-        return flow_execution.get_next_child(self)
-
-    def _execute_modify_context_value(self, flow_execution):
-        """
-        parameters - {
-            "attribute": "willpower",
-            "modifier":  {"name": "add", "args": ["$bonus"]}
-        }
-        """
-        object_pk = flow_execution.get_variable(self.variable_name)
-        if object_pk is None:
-            raise RuntimeError(
-                f"Flow variable '{self.variable_name}' is undefined - "
-                "cannot modify context value."
-            )
-
-        attribute_name = self.parameters["attribute"]
-
-        # resolve modifier makes a partial function based on name and value
-        modifier_callable = self.resolve_modifier(flow_execution)
-
-        flow_execution.context.modify_context_value(
-            key=object_pk,
-            attribute=attribute_name,
-            modifier=modifier_callable,
-        )
-        return flow_execution.get_next_child(self)
-
-    def resolve_modifier(self, flow_execution) -> callable:
-        """
-        Resolves a modifier into a callable function for context modification.
-
-        The ONLY valid schema for the modifier is:
-
-            {
-                "name": <str>,         # REQUIRED: the operator name (e.g. "add", "mul")
-                "args": <list>,        # OPTIONAL: positional arguments
-                "kwargs": <dict>       # OPTIONAL: keyword arguments
-            }
-
-        - Any deviation from this schema (missing "name", extra keys, wrong types)
-        raises ValueError.
-        - Operator name must be one of: add, sub, mul, truediv, floordiv, mod, pow,
-        neg, pos, abs, eq, ne, lt, le, gt, ge.
-        - Arguments that are strings starting with "$" will be resolved from flow
-        variables, supporting dot notation for attribute access (e.g., "$foo.bar").
-
-        Args:
-            flow_execution: The FlowExecution instance for resolving variables.
-
-        Returns:
-            Callable that can be used as a modifier.
-
-        Example:
-            modifier = {
-                "name": "add",
-                "args": [3]
-            }
-            # Will produce: lambda old_value: operator.add(old_value, 3)
-        """
-
-        OP_FUNCS = {
-            "add": operator.add,
-            "sub": operator.sub,
-            "mul": operator.mul,
-            "truediv": operator.truediv,
-            "floordiv": operator.floordiv,
-            "mod": operator.mod,
-            "pow": operator.pow,
-            "neg": operator.neg,
-            "pos": operator.pos,
-            "abs": operator.abs,
-            "eq": operator.eq,
-            "ne": operator.ne,
-            "lt": operator.lt,
-            "le": operator.le,
-            "gt": operator.gt,
-            "ge": operator.ge,
-        }
-
-        # Only accept a dict or a JSON string that parses to a dict
-        mod_spec = self.parameters.get("modifier")
-        # if it's an integer, return partial of add with that integer
-        if isinstance(mod_spec, int):
-            return functools.partial(operator.add, mod_spec)
-        if isinstance(mod_spec, str):
-            try:
-                data = json.loads(mod_spec)
-            except Exception:
-                raise ValueError("Modifier must be a JSON object string or dict.")
-        elif isinstance(mod_spec, dict):
-            data = mod_spec
-        else:
-            raise ValueError("Modifier must be a JSON object string or dict.")
-
-        # Strict schema enforcement
-        allowed_keys = {"name", "args", "kwargs"}
-        if not isinstance(data, dict):
-            raise ValueError("Modifier must be a dict.")
-        if set(data.keys()) - allowed_keys:
-            raise ValueError(
-                f"Modifier contains unknown keys: {set(data.keys()) - allowed_keys}"
-            )
-        if "name" not in data or not isinstance(data["name"], str):
-            raise ValueError("Modifier must have a 'name' key of type str.")
-        if "args" in data and not isinstance(data["args"], list):
-            raise ValueError("Modifier 'args' must be a list if present.")
-        if "kwargs" in data and not isinstance(data["kwargs"], dict):
-            raise ValueError("Modifier 'kwargs' must be a dict if present.")
-
-        func_name = data["name"]
-        if func_name not in OP_FUNCS:
-            raise ValueError(f"Unknown modifier/operator: {func_name}")
-        func = OP_FUNCS[func_name]
-
-        args = data.get("args", [])
-        kwargs = data.get("kwargs", {})
-
-        resolved_args = [flow_execution.resolve_flow_reference(a) for a in args]
-        resolved_kwargs = {
-            k: flow_execution.resolve_flow_reference(v) for k, v in kwargs.items()
-        }
-
-        return functools.partial(func, *resolved_args, **resolved_kwargs)
-
-    def _execute_call_service_function(self, flow_execution):
-        """
-        Executes a step that calls a service function.
-        """
-        service_function = flow_execution.get_service_function(self.variable_name)
-        result = service_function(flow_execution, **self.parameters)
-        result_var = self.parameters.get("result_variable")
-        if result_var:
-            flow_execution.set_variable(result_var, result)
-        return flow_execution.get_next_child(self)
-
-    def _execute_emit_flow_event(self, flow_execution):
-        """Emit a :class:`FlowEvent` and forward it to the trigger registry.
-
-        Values in ``parameters['data']`` can reference variables from the running
-        flow using ``$var`` syntax. Each reference is resolved before the event
-        is created. An example step that emits a ``glance`` event with caller and
-        target information looks like::
-
-            FlowStepDefinition(
-                action=FlowActionChoices.EMIT_FLOW_EVENT,
-                variable_name="glance",
-                parameters={
-                    "event_type": "glance",
-                    "data": {"caller": "$caller.pk", "target": "$target.pk"},
-                },
-            )
-
-        The created event is stored in the execution context and propagated to
-        the room's trigger registry. If ``event.stop_propagation`` becomes
-        ``True`` the remaining triggers will not execute.
-        """
-        event_type = self.parameters.get("event_type", self.variable_name)
-        event_data = self.parameters.get("data", {})
-        resolved_data = {
-            key: flow_execution.resolve_flow_reference(value)
-            for key, value in event_data.items()
-        }
-        flow_event = FlowEvent(event_type, source=flow_execution, data=resolved_data)
-        flow_execution.context.store_flow_event(self.variable_name, flow_event)
-        # Propagate event to trigger registry
-        trigger_registry = flow_execution.get_trigger_registry()
-        trigger_registry.process_event(
-            flow_event, flow_execution.flow_stack, flow_execution.context
-        )
-        if flow_event.stop_propagation:
-            return None
-        return flow_execution.get_next_child(self)
-
-    def _execute_emit_flow_event_for_each(self, flow_execution):
-        """Emit a FlowEvent for every item in an iterable."""
-        iterable_ref = self.parameters.get("iterable")
-        if iterable_ref is None:
-            return flow_execution.get_next_child(self)
-
-        iterable = flow_execution.resolve_flow_reference(iterable_ref)
-        event_type = self.parameters.get("event_type", self.variable_name)
-        base_data = self.parameters.get("data", {})
-        item_key = self.parameters.get("item_key", "item")
-
-        next_step = flow_execution.get_next_child(self)
-
-        for idx, item in enumerate(iterable or []):
-            data = {
-                key: (
-                    item
-                    if val == "$item"
-                    else flow_execution.resolve_flow_reference(val)
-                )
-                for key, val in base_data.items()
-            }
-            if item_key:
-                data.setdefault(item_key, item)
-
-            flow_event = FlowEvent(event_type, source=flow_execution, data=data)
-            context_key = f"{self.variable_name}_{idx}"
-            flow_execution.context.store_flow_event(context_key, flow_event)
-            trigger_registry = flow_execution.get_trigger_registry()
-            trigger_registry.process_event(
-                flow_event, flow_execution.flow_stack, flow_execution.context
-            )
-            if flow_event.stop_propagation:
-                return None
-
-        return next_step
-
-    def __str__(self):
-        return f"{self.flow.name} - Step: {self.id} ({self.action})"
-
-
-class TriggerDefinition(models.Model):
-    """Reusable template describing when to launch another flow.
-
-    ``base_filter_condition`` allows simple equality checks against event data to
-    decide if the trigger should fire. For example, given the ``glance`` event
-    emitted in the example above::
-
-        TriggerDefinition(
-            name="on glance at me",
-            event=Event.objects.get(key="glance"),
-            flow_definition=response_flow,
-            base_filter_condition={"target": 5},
-        )
-
-    A trigger based on this definition will only activate when the ``glance``
-    event's ``target`` equals ``5`` (the object's primary key).
-    """
-
-    name = models.CharField(max_length=255, unique=True)
-    event = models.ForeignKey(
-        "Event",
-        on_delete=models.CASCADE,
-        db_column="event_id",
-        help_text="The event this trigger listens for.",
-    )
-    flow_definition = models.ForeignKey(
-        FlowDefinition,
-        on_delete=models.CASCADE,
-        help_text="The flow to execute when this trigger activates.",
-    )
-    base_filter_condition = models.JSONField(
-        blank=True,
-        null=True,
-        help_text="Base JSON condition to filter when this trigger is valid.",
-    )
-
-    def _resolve_self_placeholders(self, conditions: dict | None, obj) -> dict:
-        if not conditions:
-            return {}
-        resolved = {}
-        for key, value in conditions.items():
-            if value == "$self":
-                resolved[key] = obj
-            elif value == "$self.pk":
-                resolved[key] = getattr(obj, "pk", obj)
-            else:
-                resolved[key] = value
-        return resolved
-
-    def matches_event(self, event: "FlowEvent", obj=None) -> bool:
-        """Check if this trigger definition matches ``event``.
-
-        Args:
-            event: Event to compare against.
-            obj: Object the trigger is attached to for ``$self`` resolution.
-
-        Returns:
-            bool: True if event type matches and conditions pass.
-        """
-        conditions = self._resolve_self_placeholders(self.base_filter_condition, obj)
-        return self.event.key == event.event_type and event.matches_conditions(
-            conditions
-        )
-
-    description = models.TextField(
-        blank=True, null=True, help_text="Optional description of the trigger."
-    )
-    priority = models.IntegerField(
-        default=0,
-        help_text="Higher priority triggers fire first.",
-    )
-
-    def __str__(self):
-        return self.name
-
-
-class Trigger(models.Model):
-    """
-    Represents an active trigger on an object, based on a TriggerDefinition.
-    """
-
-    trigger_definition = models.ForeignKey(
-        TriggerDefinition,
-        on_delete=models.CASCADE,
-        help_text="The trigger template this is based on.",
-    )
-    obj = models.ForeignKey(
-        "objects.ObjectDB",
-        on_delete=models.CASCADE,
-        related_name="triggers",
-        help_text="The object this trigger is associated with.",
-    )
-    additional_filter_condition = models.JSONField(
-        blank=True,
-        null=True,
-        help_text="Optional JSON condition to further refine when this trigger activates.",
-    )
-
-    @cached_property
-    def trigger_data_items(self) -> List["TriggerData"]:
-        return list(self.data.all())
-
-    @property
-    def data_map(self) -> dict[str, str]:
-        """Returns a mapping of ``{key: value}`` for all :class:`TriggerData`."""
-        return {d.key: d.value for d in self.trigger_data_items}
-
-    @property
-    def priority(self) -> int:
-        """Delegates to the trigger definition's priority."""
-        return self.trigger_definition.priority
-
-    def should_trigger_for_event(self, event: "FlowEvent") -> bool:
-        """Determines if this trigger should fire for the given ``event``.
-
-        The following rules apply:
-        1. The trigger's definition must match the event type and pass its conditions
-        2. The trigger's additional conditions must pass (if any)
-        """
-        if not self.trigger_definition.matches_event(event, obj=self.obj):
-            return False
-
-        additional = self.trigger_definition._resolve_self_placeholders(
-            self.additional_filter_condition, self.obj
-        )
-        return event.matches_conditions(additional)
-
-    def __str__(self):
-        return f"{self.trigger_definition.name} for {self.obj.key}"
-
-
-class TriggerData(models.Model):
-    """
-    Stores long-lived, arbitrary data associated with a specific Trigger.
-    """
-
-    trigger = models.ForeignKey(
-        Trigger,
-        on_delete=models.CASCADE,
-        related_name="data",
-        help_text="The specific trigger instance this data is associated with.",
-    )
-    key = models.CharField(max_length=255, help_text="The data key.")
-    value = models.TextField(help_text="The data value.")
-
-    class Meta:
-        unique_together = ("trigger", "key")
-
-    def __str__(self):
-        return f"Data for {self.trigger} - {self.key}: {self.value}"
